refactor(infra): infra 폴더 내의 디렉토리 구조 리팩토링
- infra 폴더의 경우 특정 의존 기술이나 모든 모듈에서 공통적으로 사용하는 객체들을 모아둠
- 각 파일들이 의존하는 기술명을 기준으로 적절히 분류되도록 리팩토링
