+++ src/main/java/com/nimble/server_spring/infra/jwt/JwtAuthFilter.java
@@ -15,7 +15,7 @@ import java.io.IOException;
 
 @Component
 @RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
+public class JwtAuthFilter extends OncePerRequestFilter {
 
     private final AuthTokenProvider authTokenProvider;
