+++ src/main/java/com/nimble/server_spring/infra/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.nimble.server_spring.infra.properties;
+package com.nimble.server_spring.infra.jwt;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -6,6 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @Data
 @ConfigurationProperties(prefix = "jwt")
 public class JwtProperties {
+
     private String accessTokenSecret;
     private Integer accessTokenExpiry;
     private String refreshTokenSecret;
