+++ src/main/java/com/nimble/server_spring/infra/persistence/BooleanToYNConverter.java
@@ -1,10 +1,11 @@
-package com.nimble.server_spring.infra.jpa;
+package com.nimble.server_spring.infra.persistence;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter
 public class BooleanToYNConverter implements AttributeConverter<Boolean, String> {
+
     @Override
     public String convertToDatabaseColumn(Boolean attribute) {
         return (attribute != null && attribute) ? "Y" : "N";
