+++ src/main/java/com/nimble/server_spring/infra/web/WebFilterConfig.java
@@ -1,13 +1,14 @@
-package com.nimble.server_spring.infra.config;
+package com.nimble.server_spring.infra.web;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CommonsRequestLoggingFilter;
 import org.springframework.web.filter.CorsFilter;
 
 @Configuration
-public class CorsFilterConfig {
+public class WebFilterConfig {
 
     @Bean
     public CorsFilter corsFilter() {
@@ -22,4 +23,14 @@ public class CorsFilterConfig {
         return new CorsFilter(source);
     }
 
+    @Bean
+    public CommonsRequestLoggingFilter logFilter() {
+        CommonsRequestLoggingFilter filter = new CommonsRequestLoggingFilter();
+        filter.setIncludeQueryString(true);
+        filter.setIncludePayload(true);
+        filter.setMaxPayloadLength(10000);
+        filter.setIncludeHeaders(false);
+        filter.setAfterMessagePrefix("REQUEST DATA : ");
+        return filter;
+    }
 }
